test: DCI 패턴을 적용한 JwtTokenProviderTest 적용
